api/requirements.txt
beautifulsoup4==4.9.1
Bottleneck==1.3.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
fastai==1.0.61
fastprogress==0.2.4
Flask==1.1.2
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==1.4.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
numexpr==2.7.1
numpy==1.19.1
nvidia-ml-py3==7.352.0
opencv-python==4.3.0.36
packaging==20.4
pandas==1.1.0
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
scipy==1.5.2
six==1.15.0
soupsieve==2.0.1
torch==1.4.0
torchvision==0.5.0
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
